+++ packages/Table/src/Table.js
@@ -1,14 +1,18 @@
 /* eslint-disable react/no-array-index-key */
 import React from "react";
-
+import { v4 as uuidv4 } from "uuid";
 import { extractChildren } from "@paprika/helpers";
 import PropTypes from "prop-types";
 import * as constants from "@paprika/constants/lib/Constants";
 import ColumnDefinition from "./components/ColumnDefinition";
 import * as sc from "./Table.styles";
+import { handleBlur, handleFocus, handleKeyDown, handleClick } from "./event";
 
 export default function Table(props) {
-  const { borderType, children, hasZebraStripes, data, a11yText, ...moreProps } = props;
+  const { borderType, children, hasZebraStripes, data, a11yText, enableArrowKeyNavigation, ...moreProps } = props;
+  const [tableId] = React.useState(() => `table_${uuidv4()}`);
+
+  const refFocus = React.useRef(null);
 
   const { "Table.ColumnDefinition": extractedColumnDefinitions } = extractChildren(children, [
     "Table.ColumnDefinition",
@@ -21,8 +25,23 @@ export default function Table(props) {
     ColumnDefinitions = [extractedColumnDefinitions];
   }
 
+  const qty = {
+    columnsLength: extractedColumnDefinitions.length,
+    rowsLength: data.length,
+  };
+
+  const arrowKeyNavigationProps = enableArrowKeyNavigation
+    ? {
+        onFocus: handleFocus({ refFocus, tableId }),
+        onBlur: handleBlur({ refFocus, tableId }),
+        onKeyDown: handleKeyDown({ refFocus, tableId, ...qty }),
+        onClick: handleClick({ refFocus, tableId }),
+        tabIndex: -1,
+      }
+    : {};
+
   return (
-    <sc.Table aria-label={a11yText} {...moreProps}>
+    <sc.Table {...(enableArrowKeyNavigation ? { role: "grid" } : {})} aria-label={a11yText} id={tableId} {...moreProps}>
       <sc.Thead>
         <tr>
           {ColumnDefinitions.map((columnDefinition, columnIndex) => {
@@ -50,17 +69,33 @@ export default function Table(props) {
           return (
             <tr key={rowIndex}>
               {ColumnDefinitions.map((columnDefinition, columnIndex) => {
+                const position = { "data-row-index": rowIndex, "data-column-index": columnIndex };
+
                 const { cell, header, ...moreColumnProps } = columnDefinition.props;
 
                 if (typeof cell === "function")
                   return (
-                    <sc.TD borderType={borderType} key={columnIndex} {...moreColumnProps}>
+                    <sc.TD
+                      borderType={borderType}
+                      key={columnIndex}
+                      {...moreColumnProps}
+                      {...position}
+                      {...arrowKeyNavigationProps}
+                      {...(columnIndex === 0 && rowIndex === 0 ? { tabIndex: 0 } : {})}
+                    >
                       {cell({ row, rowIndex, columnIndex })}
                     </sc.TD>
                   );
                 if (typeof cell === "string")
                   return (
-                    <sc.TD borderType={borderType} key={columnIndex} {...moreColumnProps}>
+                    <sc.TD
+                      borderType={borderType}
+                      key={columnIndex}
+                      {...moreColumnProps}
+                      {...position}
+                      {...arrowKeyNavigationProps}
+                      {...(columnIndex === 0 && rowIndex === 0 ? { tabIndex: 0 } : {})}
+                    >
                       {typeof row[cell] !== "undefined" ? row[cell] : `Error: ${cell} doesn't exist`}
                     </sc.TD>
                   );
@@ -94,12 +129,15 @@ const propTypes = {
   hasZebraStripes: PropTypes.bool,
   /** Array of data to be stored in the Table */
   data: PropTypes.arrayOf(PropTypes.shape({})),
+  /** For authors use only, use case: inline editing. */
+  enableArrowKeyNavigation: PropTypes.bool,
 };
 
 const defaultProps = {
   borderType: Table.types.border.HORIZONTAL,
   data: [],
   hasZebraStripes: false,
+  enableArrowKeyNavigation: false,
 };
 
 Table.displayName = "Table";