+++ packages/Filter/CHANGELOG.md
@@ -5,3 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+
+## [1.0.2-alpha.0] - 2021-01-27
+
+### Changed
+
+- Changed `<Select />` to `<ListBox />` everywhere in the component.
+
+## [1.1.2-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.