+++ packages/ListBox/CHANGELOG.md
@@ -60,3 +60,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed the potential mapping children error.
+
+## [1.3.3-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.