+++ packages/Popover/CHANGELOG.md
@@ -24,3 +24,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.0.4-alpha.0] - 2020-12-16
 
 - Pass the `Event` object down for `<Popover.Trigger />` `onClick`.
+
+## [1.0.7-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.