+++ packages/Sortable/CHANGELOG.md
@@ -9,3 +9,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## [1.0.2-alpha.2] - 2020-11-24
 
 - Added `isDragDisabled` and `handleElement` prop on `Sortable.Item`.
+
+## [1.1.5-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.