+++ packages/Uploader/CHANGELOG.md
@@ -51,3 +51,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Fixed
 
 - Fixed the potential mapping children error.
+
+## [2.0.4-alpha.0] - 2021-02-01
+
+### Changed
+
+- Replace `nanoid` by `uuid`.